requirements.txt
Flask==0.11.1
cf-deployment-tracker==1.0.4
cloudant==2.4.0
watson-developer-cloud==0.25.2
pysolr==3.6.0
PyYAML>=3.11
